begin chainSync tests
