+++ dot/sync/chain_sync_test.go
@@ -1,0 +1,48 @@
+package sync
+
+import (
+	"math/big"
+	"testing"
+
+	syncmocks "github.com/ChainSafe/gossamer/dot/sync/mocks"
+	"github.com/ChainSafe/gossamer/dot/types"
+	"github.com/ChainSafe/gossamer/lib/common"
+	"github.com/ChainSafe/gossamer/lib/trie"
+
+	"github.com/libp2p/go-libp2p-core/peer"
+
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func newTestChainSync(t *testing.T) (*chainSync, <-chan *types.BlockData) {
+	header, err := types.NewHeader(common.NewHash([]byte{0}), trie.EmptyHash, trie.EmptyHash, big.NewInt(0), types.Digest{})
+	require.NoError(t, err)
+
+	bs := new(syncmocks.MockBlockState)
+	bs.On("BestBlockHeader").Return(header, nil)
+
+	net := new(syncmocks.MockNetwork)
+	net.On("DoBlockRequest", mock.AnythingOfType("peer.ID"), mock.AnythingOfType("*BlockRequestMessage")).Return(nil, nil)
+
+	readyBlocks := make(chan *types.BlockData, MAX_RESPONSE_SIZE)
+	return newChainSync(bs, net, readyBlocks), readyBlocks
+}
+
+func TestChainSync_SetPeerHead(t *testing.T) {
+	cs, _ := newTestChainSync(t)
+
+	testPeer := peer.ID("noot")
+	hash := common.Hash{0xa, 0xb}
+	number := big.NewInt(1000)
+	err := cs.setPeerHead(testPeer, hash, number)
+	require.NoError(t, err)
+
+	expected := &peerState{
+		hash:   hash,
+		number: number,
+	}
+	require.Equal(t, expected, cs.peerState[testPeer])
+	require.Equal(t, expected, <-cs.workQueue)
+	require.True(t, cs.pendingBlocks.hasBlock(hash))
+}
